PTSD-1410 fixed expectedFrameRate to be floating point, not integer

+++ pkg/web/models.go
@@ -191,7 +191,7 @@ type SocketBasecallerObject struct {
 
 	// The expected (not measured) canonical frame rate
 	// Example: 100
-	ExpectedFrameRate int32 `json:"expectedFrameRate"`
+	ExpectedFrameRate float64 `json:"expectedFrameRate"`
 
 	// The inversion of photoelectron gain of the sensor pixels.
 	// Example: 1.4
